+++ src/shapes/rect-host.ts
@@ -138,6 +138,28 @@ export class RectHost extends ShapeHost {
   }
   //#endregion
 
+  //#region move shape
+  public override moveShapeUp(amount: number): void {
+    const coords = this.getEditPointsCoordsFromForm().map((c) => ({ ...c, y: this.toFixed(c.y - amount) }));
+    this.updateFormWithCoords(coords);
+  }
+
+  public override moveShapeRight(amount: number): void {
+    const coords = this.getEditPointsCoordsFromForm().map((c) => ({ ...c, x: this.toFixed(c.x + amount) }));
+    this.updateFormWithCoords(coords);
+  }
+
+  public override moveShapeDown(amount: number): void {
+    const coords = this.getEditPointsCoordsFromForm().map((c) => ({ ...c, y: this.toFixed(c.y + amount) }));
+    this.updateFormWithCoords(coords);
+  }
+
+  public override moveShapeLeft(amount: number): void {
+    const coords = this.getEditPointsCoordsFromForm().map((c) => ({ ...c, x: this.toFixed(c.x - amount) }));
+    this.updateFormWithCoords(coords);
+  }
+  //#endregion
+
   //#region import
   public override loadFromElement(svg: SVGRectElement) {
     this.svg = svg;